Interference mutlinomial distribution

Interference mutlinomial distribution

interf_dmultinom(x, prob, k, interf)

Arguments

x
Vector of length equal to `data` of integers in `0:sum(data)`.

prob
Numeric non-negative vector of length `length(data)`, specifying the probability for classes; is internally normalized to sum 1. Infinite and missing values are not allowed.

k
Interference parameters. See details.

interf
Interference relationships. Output of interf_pair_pos

Value

Returns probability density vector of length equal to `length(prob)`

Details

Interference parameters describe how the probability density of `dmultionom(x, prob=prob)` is altered depending on which is the first draw from the mutlinomial distribution.

Two types of interference parameter can be specified. For example:

Modulators, which should be passed as a named vector, with names "k<n>":

`k = c("k0"=0.5)` - The probability of choosing the same event as the first drawn event will be mutliplied by 0.5

`k = c("k1"=0.5,"k2"=2)` - The probability of choosing the first element in x again if it was drawn first will be multiplied by 0.5, whereas if the second element was drawn first then the probability of picking it again is doubled.

Interferences, passed as a named vector, with names "k_<n>_<m>":

`k = c("k_1_2" = 0.5, "k_1_3" = 2, "k_2_3")` - The probability of choosing event 1 if event 2 drawn first is half, and vice versa. The probability of drawing event 1 if event 3 was drawn first is doubled, and vice versa. The probability of drawing event 2 if event 3 was drawn first is doubled, and vice versa.

Interferences are thus symmetrical between events.